+++ myserver/include/protocol/https/https.h
@@ -63,7 +63,7 @@ public:
   }
 
 	virtual int controlConnection (ConnectionPtr a, char *b1, char *b2,
-                                int bs1, int bs2, u_long nbtr, u_long id)
+                                u_long bs1, u_long bs2, u_long nbtr, u_long id)
   {
     int ret = 0;
     Https https;
